Refactored healthcheck to work for tests too.
