+++ partcad/src/partcad/__init__.py
@@ -41,7 +41,7 @@ from .shape import Shape
 from .user_config import user_config
 from .logging_ansi_terminal import init as logging_ansi_terminal_init
 from .logging_ansi_terminal import fini as logging_ansi_terminal_fini
-from .healthcheck import healthcheck
+from . import healthcheck
 from . import logging
 from . import utils
 from . import exception
