+++ partcad/src/partcad/healthcheck/__init__.py
@@ -1,4 +1,4 @@
-from .healthcheck import run_tests
+from .tests import run_tests
 
 __all__ = [
     "run_tests",
